migrate to Spring boot - wip - json numbers

+++ server/src/main/scala/kpn/api/common/Check.scala
@@ -3,6 +3,6 @@ package kpn.api.common
 case class Check(
   nodeId: Long,
   nodeName: String,
-  actual: Int,
-  expected: Int
+  actual: Long,
+  expected: Long
 )
