+++ server/src/main/scala/kpn/api/common/FactCount.scala
@@ -2,4 +2,4 @@ package kpn.api.common
 
 import kpn.api.custom.Fact
 
-case class FactCount(fact: Fact, count: Int)
+case class FactCount(fact: Fact, count: Long)
