+++ server/src/main/scala/kpn/api/common/LocationCandidate.scala
@@ -1,3 +1,3 @@
 package kpn.api.common
 
-case class LocationCandidate(location: Location, percentage: Int)
+case class LocationCandidate(location: Location, percentage: Long)
