+++ server/src/main/scala/kpn/api/common/NetworkIntegrityCheck.scala
@@ -1,3 +1,3 @@
 package kpn.api.common
 
-case class NetworkIntegrityCheck(count: Int, failed: Int)
+case class NetworkIntegrityCheck(count: Long, failed: Long)
