+++ server/src/main/scala/kpn/api/common/NodeIntegrityCheck.scala
@@ -1,3 +1,3 @@
 package kpn.api.common
 
-case class NodeIntegrityCheck(nodeName: String, nodeId: Long, actual: Int, expected: Int, failed: Boolean)
+case class NodeIntegrityCheck(nodeName: String, nodeId: Long, actual: Long, expected: Long, failed: Boolean)
