+++ server/src/main/scala/kpn/api/common/changes/details/ChangeKey.scala
@@ -3,7 +3,7 @@ package kpn.api.common.changes.details
 import kpn.api.custom.Timestamp
 
 case class ChangeKey(
-  replicationNumber: Int,
+  replicationNumber: Long,
   timestamp: Timestamp,
   changeSetId: Long,
   elementId: Long
