+++ server/src/main/scala/kpn/api/common/changes/filter/ChangesFilterPeriod.scala
@@ -2,8 +2,8 @@ package kpn.api.common.changes.filter
 
 case class ChangesFilterPeriod(
   name: String,
-  totalCount: Int,
-  impactedCount: Int,
+  totalCount: Long,
+  impactedCount: Long,
   current: Boolean = false,
   selected: Boolean = false,
   periods: Seq[ChangesFilterPeriod] = Seq.empty
