+++ server/src/main/scala/kpn/api/common/data/Meta.scala
@@ -3,7 +3,7 @@ package kpn.api.common.data
 import kpn.api.custom.Timestamp
 
 trait Meta {
-  def version: Int
+  def version: Long
 
   def changeSetId: Long
