+++ server/src/main/scala/kpn/api/common/data/MetaData.scala
@@ -2,4 +2,4 @@ package kpn.api.common.data
 
 import kpn.api.custom.Timestamp
 
-case class MetaData(version: Int, timestamp: Timestamp, changeSetId: Long) extends Meta
+case class MetaData(version: Long, timestamp: Timestamp, changeSetId: Long) extends Meta
