+++ server/src/main/scala/kpn/api/common/diff/node/NodeMoved.scala
@@ -2,4 +2,4 @@ package kpn.api.common.diff.node
 
 import kpn.api.common.LatLonImpl
 
-case class NodeMoved(before: LatLonImpl, after: LatLonImpl, distance: Int)
+case class NodeMoved(before: LatLonImpl, after: LatLonImpl, distance: Long)
