+++ server/src/main/scala/kpn/api/common/network/NetworkRouteInfo.scala
@@ -6,8 +6,8 @@ import kpn.api.custom.Timestamp
 case class NetworkRouteInfo(
   id: Long,
   name: String,
-  wayCount: Int,
-  length: Int, // length in meter
+  wayCount: Long,
+  length: Long, // length in meter
   role: Option[String],
   relationLastUpdated: Timestamp,
   lastUpdated: Timestamp,
