+++ server/src/main/scala/kpn/api/common/node/NodeChangesPage.scala
@@ -8,6 +8,6 @@ case class NodeChangesPage(
   filter: ChangesFilter,
   changes: Seq[NodeChangeInfo],
   incompleteWarning: Boolean,
-  totalCount: Int,
-  changeCount: Int
+  totalCount: Long,
+  changeCount: Long
 )
