+++ server/src/main/scala/kpn/api/common/node/NodeDetailsPage.scala
@@ -5,5 +5,5 @@ import kpn.api.common.NodeInfo
 case class NodeDetailsPage(
   nodeInfo: NodeInfo,
   references: NodeReferences,
-  changeCount: Int
+  changeCount: Long
 )
