+++ server/src/main/scala/kpn/api/common/planner/RouteLegFragment.scala
@@ -3,7 +3,7 @@ package kpn.api.common.planner
 case class RouteLegFragment(
   lat: String,
   lon: String,
-  meters: Int,
-  orientation: Int,
-  streetIndex: Option[Int]
+  meters: Long,
+  orientation: Long,
+  streetIndex: Option[Long]
 )
