+++ server/src/main/scala/kpn/api/common/planner/RouteLegRoute.scala
@@ -3,7 +3,7 @@ package kpn.api.common.planner
 case class RouteLegRoute(
   source: RouteLegNode,
   sink: RouteLegNode,
-  meters: Int,
+  meters: Long,
   segments: Seq[RouteLegSegment],
   streets: Seq[String]
 )
