+++ server/src/main/scala/kpn/api/common/route/RouteDetailsPage.scala
@@ -3,5 +3,5 @@ package kpn.api.common.route
 case class RouteDetailsPage(
   route: RouteInfo,
   references: RouteReferences,
-  changeCount: Int
+  changeCount: Long
 )
