+++ server/src/main/scala/kpn/api/common/route/RouteInfo.scala
@@ -12,7 +12,7 @@ case class RouteInfo(
   summary: RouteSummary,
   active: Boolean,
   orphan: Boolean,
-  version: Int,
+  version: Long,
   changeSetId: Long,
   lastUpdated: Timestamp,
   tags: Tags,
