+++ server/src/main/scala/kpn/api/common/route/RouteMapPage.scala
@@ -2,5 +2,5 @@ package kpn.api.common.route
 
 case class RouteMapPage(
   route: RouteInfo,
-  changeCount: Int
+  changeCount: Long
 )
