+++ server/src/main/scala/kpn/api/common/subset/SubsetInfo.scala
@@ -3,9 +3,9 @@ package kpn.api.common.subset
 case class SubsetInfo(
   country: String,
   networkType: String,
-  networkCount: Int = 0,
-  factCount: Int = 0,
-  changesCount: Int = 0,
-  orphanNodeCount: Int = 0,
-  orphanRouteCount: Int = 0
+  networkCount: Long = 0,
+  factCount: Long = 0,
+  changesCount: Long = 0,
+  orphanNodeCount: Long = 0,
+  orphanRouteCount: Long = 0
 )
