+++ server/src/main/scala/kpn/api/common/tiles/ClientPoiDefinition.scala
@@ -1,3 +1,3 @@
 package kpn.api.common.tiles
 
-case class ClientPoiDefinition(name: String, icon: String, minLevel: Int, defaultLevel: Int)
+case class ClientPoiDefinition(name: String, icon: String, minLevel: Long, defaultLevel: Long)
