+++ server/src/main/scala/kpn/api/custom/ApiResponse.scala
@@ -1,3 +1,3 @@
 package kpn.api.custom
 
-case class ApiResponse[T](situationOn: Option[Timestamp], version: Int, result: Option[T])
+case class ApiResponse[T](situationOn: Option[Timestamp], version: Long, result: Option[T])
