+++ server/src/main/scala/kpn/core/db/RouteDocViewResult.scala
@@ -1,3 +1,3 @@
 package kpn.core.db
 
-case class RouteDocViewResult(total_rows: Int, offset: Int, rows: Seq[RouteDocViewResultRow])
+case class RouteDocViewResult(total_rows: Long, offset: Long, rows: Seq[RouteDocViewResultRow])
