+++ server/src/main/scala/kpn/core/db/couch/ViewResult2.scala
@@ -1,3 +1,3 @@
 package kpn.core.db.couch
 
-case class ViewResult2(total_rows: Int, offset: Option[Int], rows: Seq[ViewResultRow2])
+case class ViewResult2(total_rows: Long, offset: Option[Long], rows: Seq[ViewResultRow2])
