+++ server/src/main/scala/kpn/server/analyzer/engine/analysis/route/analyzers/RouteMapAnalyzer.scala
@@ -81,7 +81,7 @@ class RouteMapAnalyzer(context: RouteAnalysisContext) {
   private def toTrackSegment(segment: Segment): TrackSegment = {
 
     val fragments: Seq[TrackSegmentFragment] = segment.fragments.flatMap { fragment =>
-      val streetIndex: Option[Int] = fragment.fragment.way.tags("name") match {
+      val streetIndex: Option[Long] = fragment.fragment.way.tags("name") match {
         case None => None
         case Some(street) =>
           context.streets match {
