+++ server/src/main/scala/kpn/server/analyzer/engine/analysis/route/segment/SegmentBuilder.scala
@@ -60,7 +60,7 @@ class SegmentBuilder {
     }
   }
 
-  private def length(sfs: Seq[SegmentFragment]): Int = {
+  private def length(sfs: Seq[SegmentFragment]): Long = {
     sfs.map(_.fragment.meters).sum
   }
