+++ server/src/main/scala/kpn/server/analyzer/engine/changes/changes/RelationAnalyzer.scala
@@ -38,6 +38,6 @@ trait RelationAnalyzer {
 
   def lastUpdated(relation: Relation): Timestamp
 
-  def waysLength(relation: Relation): Int
+  def waysLength(relation: Relation): Long
 
 }
