+++ server/src/test/scala/kpn/server/json/JsonExample.scala
@@ -1,3 +1,3 @@
 package kpn.server.json
 
-case class JsonExample(name: String, age: Int)
+case class JsonExample(name: String, age: Long)
